chore: rename grids stories
